# argpcp: reduce re-allocations of intermediate matrices