@schmidmt 2024-05-31
@sd2k If it would be helpful to create a new release, let me know. I'd be happy to release a patch version. Otherwise, I'll wrap it up with a larger update later.