@sd2k 2024-05-31
@schmidmt a new release in the next few weeks would be great but it's not super urgent, I can use a patch in my Cargo.toml for now 👍