@sd2k 2024-07-03
Hi @schmidmt! Do you have bandwidth to create a new patch version with this fix in? Thanks!